// provider/enum/filter-type.enum.ts
export enum EfilterType {
  USER_STATUS = 'USER_STATUS',
  INSPECTION_STATUS = 'INSPECTION_STATUS',
  UPI_STATUS = 'UPI_STATUS',
  CREDIT_STATUS = 'CREDIT_STATUS',
  TOKEN_STATUS = 'TOKEN_STATUS',
  REQUEST_STATUS = 'REQUEST_STATUS',
  APPLICATION_STATE = 'APPLICATION_STATE',
  REPORT = 'REPORT',
  NO_SITEVISIT = 'NO_SITEVISIT',
  WITHDRAW = 'WITHDRAW'
}